% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMark.R
\name{add/clear MarkLine}
\alias{add/clear MarkLine}
\alias{addMarkLine}
\title{Mark Line}
\usage{
addMarkLine(chart, mpdata, series = 1, ...)
}
\examples{
mpdata = list(list(coord=c("2016-01-12",10)), list(coord=c("2016-01-17",15)))
addMarkPoint(k1,mpdata)
mldf1 = data.frame(x0=c("2016-01-02","2016-01-17"),y0=c(30,30), x=c("2016-01-05","2016-01-20"),y=c(20,40), color=c("steelblue","orange"), symbol="pin", symbolRotate=c(0,180), width=1,type="solid")
addMarkPoint(k1,mpdf1)
}

