% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addMark.R
\name{add/clear MarkPoints}
\alias{add/clear MarkPoints}
\alias{addMarkPoint}
\title{Mark Point}
\usage{
addMarkPoint(chart, mpdata, series = 1, ...)
}
\examples{
mpdata = list(list(coord=c("2016-01-12",10)), list(coord=c("2016-01-17",15)))
addMarkPoint(k1,mpdata)
mpdf1 = data.frame(x=c("2016-01-02","2016-01-20"),y=c(30,30), color=c("steelblue","orange"), symbol="pin", symbolRotate=c(0,180), ...)
addMarkPoint(k1,mpdf1)
addMarkPoint(k1) <=> clearMarkPoint
}

