% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addMarkLine}
\alias{addMarkLine}
\title{Add \code{markLine} to An Echarts Object}
\usage{
addMarkLine(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}
Short form: \cr
\tabular{llll}{
 [col 1] series name|index \tab [col 2] line name \tab [col 3] Line type
 \tab [col 4] Light effect \cr
 String or number \tab String|NA \tab min|max|average|lm \tab TRUE|FALSE
}
Full form: \cr
\tabular{llllllll}{
 [col 1] series name|index \tab [col 2] line name \tab [col 3] Value
 \tab [col 4] P0 x \tab [col 5] P0 y \tab [col 6] P1 x \tab [col 7] P1 y
 \tab [col 8] Light effect \cr
 String|number \tab String|NA \tab num \tab x val \tab y val \tab x val
 \tab y val \tab TRUE|FALSE
}
Examples:
\itemize{
 \item E.g., both \code{t(c('male', NA, 'average', FALSE))} and \code{t(c(1, NA, 'average', FALSE))}
 refer to an average markline of the series 'male', only available for line,
 linesmooth, bar, scatter, bubble charts. 'lm' refers to linear regresson
 markline which is only available for scatters and bubbles.
 \code{t(c(1, NA, 'average', TRUE))} opens light effects of series 'male'. \cr
 \item In line, bar, k and scatter charts, 'P0 x','P0 y','P1 x','P1 y' are
 comprehended as coordinates. In map charts, these coordinates should be lattitudes
 and longitudes. \code{t(c('male', NA, 100, 0, 5, 100, 5, TRUE))} opens light effects of series 'male'.
}}
}
\description{
Add \code{markLine} to An Echarts Object
}

