% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{addMarkPoint}
\alias{addMarkPoint}
\title{Add \code{markPoint} to An Echarts Object}
\usage{
addMarkPoint(chart, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}
Show markpoints, default to NULL.
The grammar is a data.frame or matrix with 4 or 6 columns:
Short form: \cr
\tabular{llll}{
 <col 1> series name|index \tab <col 2> Point name \tab <col 3> Point type \tab <col 4> Light effect \cr
 String or number \tab String|NA \tab min|max \tab TRUE|FALSE
 }
Full form: \cr
\tabular{llllll}{
 [col 1] series name|index \tab [col 2] Point name \tab [col 3] Value \tab [col 4] P x
 \tab [col 5] P y \tab [col 6] Light effect \cr
 String or number \tab String|NA \tab num \tab x val \tab y val \tab TRUE|FALSE
}
Examples: \cr
\itemize{
 \item E.g., both \code{t(c('male',NA,'min',FALSE))} and \code{t(c(1,NA,'min',FALSE))} refer
 to a min markpoint of the series 'male', only available for line, linesmooth, bar,
 scatter, bubble charts. \code{t(c('male',NA,'min',TRUE))} opens light effects of series 'male'.
 \item In line, bar, k and scatter charts, 'P x','P y',... are comprehended
 as coordinates. In map charts, these coordinates should be lattitudes and longitudes.
 \code{t(c('male',NA,100,0,5,TRUE))} opens light effects of series 'male'.
}}
}
\description{
Add \code{markPoint} to An Echarts Object
}

