% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recharts.R
\name{&.echarts}
\alias{&.echarts}
\title{Merge the two ECharts into one output .}
\usage{
\method{&}{echarts}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{recharts}}

\item{e2}{An object of class \code{recharts}}
}
\description{
Merge the two ECharts into one output .
}
\seealso{
\code{\link{set}}
}
