% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChinaGDP}
\alias{ChinaGDP}
\title{Province-specific GDP 2012-2014 in China}
\format{A data.frame (93 * 3) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 Year \tab num \tab Years (2012, 2013, 2014, ...) \cr
 Prov \tab chr \tab Province names (Guangdong, Beijing, Shanghai, ...) \cr
 GDP \tab num \tab GDP in RMB (1103605, 1059587, 967419, ...)
}}
\description{
A data.frame comprising of province-specific GDP (2012-14) in China.
}
\examples{
data(ChinaGDP)
str(ChinaGDP)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_31_Map.html}
}

