% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chinapm25}
\alias{chinapm25}
\title{[Dataset] PM2.5 in China Cities (Fictious Data)}
\format{A data.frame (199 * 4) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 City \tab chr \tab City names (Haimen, Ordos, Zhaoyuan, ...) \cr
 Value \tab num \tab PM2.5 values (9, 12, 12, ...) \cr
 Lng \tab num \tab Longitude (121, 110, 120, 122, ...) \cr
 Lat \tab num \tab Latitude (31.9, 39.6, 37.4, 30, ...)
}}
\description{
A data.frame comprising of PM2.5 data in China cities.
}
\examples{
data(chinapm25)
str(chinapm25)
}
\references{
\url{https://madlogos.github.io/rechartsX/Basic_Plots_31_Map.html}
}
