% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{dataZoom}
\alias{dataZoom}
\title{Create a dataZoom widget for a chart}
\usage{
dataZoom(chart, show = TRUE, orient = c("horizontal", "vertical"),
  start = 0, end = 100, showDetail = TRUE, realtime = TRUE,
  zoomLock = FALSE)
}
\arguments{
\item{zoomLock}{if lock the data window width.}
}
\description{
Add the dataZoom widget
}

