% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deutsch}
\alias{deutsch}
\title{[Dataset] Deutsch Soccer Team And Clubs}
\format{A data.frame (26 * 5) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 player \tab chr \tab Player names \cr
 club \tab chr \tab Club of the players \cr
 weight \tab num \tab The weight of the connection between clubs and players \cr
 role \tab chr \tab Role of the players ('Fw', 'Mf', 'Gk', 'Df', ...) \cr
 year \tab int \tab Year tag of the Deutsch soccer team
}}
\description{
A data.frame comprising of Deutsch soccer team players and their clubs. It contains
two Deutsch team (\code{year} == 2014 and \code{year} == 2016)
}
\examples{
data(deutsch)
str(deutsch)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_12_Chord.html}
}
