% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eBmap.R
\name{eBmap}
\alias{eBmap}
\title{Create an HTML map charts widget using the ECharts(version:3.2.2) library}
\usage{
eBmap(center = c(104.114129, 37.550339), zoom = 5, roam = TRUE,
  mapStyle = NULL, title = NULL, subtitle = NULL,
  title.x = "center", title.y = "top", size = NULL)
}
\arguments{
\item{center}{center of echarts map(lon,lat)}

\item{zoom}{zoom of the map}

\item{roam}{allow roam, default is true}

\item{mapStyle}{a list object to set map style}

\item{title, subtitle}{an overall title for the plot. you can modify title widget after chart has been
generated: Echart + eTitle(title = "your title.")}

\item{title.x, title.y}{the xy coordinates of main title, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify title widget after chart has been generated: 
\code{Echart + eTitle(title="main title", x = "left", y=10)}}
}
\description{
This function creates an HTML widget to display data.frame and 
factor array, using the JavaScript library ECharts3.
}
\examples{
eBmap()
}
