% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartsExample.R
\name{echartsExample}
\alias{echartsExample}
\title{Create Echarts from Echarts Glaery}
\usage{
echartsExample = function(url = NULL,
   JScontent = NULL,
   width = NULL, height = NULL)
}
\arguments{
\item{url}{Links from 'https://ecomfe.github.io/echarts-examples/public/index.html'
(Chinese Version: 'http://echarts.baidu.com/examples.html').}

\item{JScontent}{JScodes for Echarts. An object named option should be created in these codes.}
}
\examples{
# Create from link
echartsExample("http://echarts.baidu.com/demo.html#effectScatter-bmap")

# Create from JS codes
JScodes = "
var option = {
  title: {
    text: 'ECharts entry example'
  },
  tooltip: {},
  legend: {
    data:['Sales']
  },
  xAxis: {
    data: ['shirt','cardign','chiffon shirt','pants','heels','socks']
  },
  yAxis: {},
  series: [{
    name: 'Sales',
    type: 'bar',
    data: [5, 20, 36, 10, 10, 20]
  }]
};
"
echartsExample(JScontent=JScodes)
}
