% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recharts.R
\name{echartsLayout}
\alias{echartsLayout}
\title{set multiple echarts layout}
\usage{
echartsLayout(multiEcharts)
}
\arguments{
\item{multiEcharts}{A multiple echarts object to set the layout.}
}
\description{
Use the same layout orgnization as original grDevice layout function.
}
