% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eFunnel.R
\name{eFunnel}
\alias{eFunnel}
\title{Funnel Charts}
\usage{
eFunnel(dat, fix = T, opt = list())
}
\arguments{
\item{dat}{vector, names will be treated as labels}

\item{fix}{fix the max of data into 100%, default is true.}

\item{opt}{option of ECharts.}
}
\value{
an htmlwidgets object which can be rendered by Browser.
}
\description{
ECharts style Funnel charts.
}
\examples{
x = (1:5)*20
names(x) = c("A","B","C","D","E")
eFunnel(x,theme='macarons') +eTitle(title = "Funnel Plot")
}

