% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eMarkLine.R
\name{eHLine}
\alias{eHLine}
\title{Create an HTML bar charts widget using the ECharts(version:3.2.2) library}
\usage{
eHLine(chart, yvalue, lineType = "solid", symbol = "none",
  lineWidth = 1, lineColor = "red", name = "", precision = NULL,
  seriesIndex = NULL)
}
\arguments{
\item{chart}{a recharts object}

\item{yvalue, xvalue}{x,y value for horizontal or vertical line,can be one of 
('min','max','average')}

\item{lineType}{default is solid, can be of ('solid','dashed','dotted')}

\item{symbol}{symbol type at the two ends of the mark line, none  includes 
'circle', 'rect', 'roundRect', 'triangle', 'diamond', 'pin', 'arrow'}

\item{lineWidth}{line width, default is 1}

\item{lineColor}{default is NULL,}

\item{name}{default is NULL,}

\item{precision}{default is 2,}

\item{seriesIndex}{default is NULL,}

\item{startPoint, endPoint}{start and end point for XYline}

\item{series}{an "formula" object: Associates the levels of variable
with symbol color, e.g. \code{series = ~groupName}}
}
\description{
This function creates an HTML widget to display matrix, data.frame and 
factor array, using the JavaScript library ECharts3.
}
\examples{
chart = ePoints(iris[, 3:5], theme = 2)
eHLine(chart,1,lineColor = 'blue')
eHLine(chart,'min',lineWidth = 5)
eVLine(chart,1,lineColor = 'blue')
eVLine(chart,'min',lineWidth = 5)
eMarkLine(chart, c(1,0), c(7,2.5))
}
