% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eMarkPoint.R
\name{eMarkPoint}
\alias{eMarkPoint}
\title{Create an markpoint using the ECharts(version:3.2.2) library}
\usage{
eMarkPoint(chart, dat = NULL, xvar = NULL, yvar = NULL, series = NULL,
  type = NULL, valueIndex = 0, seriesIndex = NULL, symbol = "pin",
  symbolSize = 50, symbolRotate = NULL, symbolOffset = c(0, 0),
  silent = FALSE, label = NULL, itemStyle = NULL)
}
\arguments{
\item{chart}{a recharts object}

\item{dat}{a data object (a matrix, a data frame or a factor array)}

\item{xvar, yvar}{objects of class "formula" (or one that can be coerced 
to that class):  x,y coordinates of the given data.frame colnames, e.g. 
\code{xvar = ~xAxisName}; \code{yvar = ~yAxisName}. xvar, yvar only needed for the 
data.frame data input.}

\item{series}{an "formula" object: Associates the levels of variable
with symbol color, e.g. \code{series = ~groupName}}

\item{type}{default is NULL,}

\item{valueIndex}{default is 0,}

\item{seriesIndex}{default is NULL,}

\item{symbol}{default is 'pin',}

\item{symbolSize}{default is  50,}

\item{symbolRotate}{default is NULL,}

\item{symbolOffset}{default is c(0,0),}

\item{silent}{default is FALSE,}

\item{label}{default is NULL,}

\item{itemStyle}{default is NULL}

\item{size}{an array of html widget width and height(either numeric pixels 
or percentage could be accepted): e.g. size = c(1024, 768).}
}
\description{
This function creates an HTML widget to display matrix, data.frame and 
factor array, using the JavaScript library ECharts3.
}
\examples{
dat <- cut(rnorm(1000), -4:4)
chart = eLine(dat)
eMarkPoint(chart, type=c('max','min'),valueIndex = 1)
eMarkPoint(chart, dat = data.frame(2:3,rep(200,2)),
           type=c('max','min'),valueIndex = 1)
}
