% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eParallel.R
\name{eParallel}
\alias{eParallel}
\title{Create an HTML parallel charts widget using the ECharts(version:3.2.2) library}
\usage{
eParallel(dat, series = NULL, axisList = list(), size = NULL,
  colorStyle = "black", layout = "horizontal", left = "5\%",
  right = "18\%", bottom = 100, top = 100, theme = "default",
  title = NULL, subtitle = NULL, title.x = "center", title.y = "top",
  legend = TRUE, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", axis.scale = TRUE, axis.nameGap = 20,
  axis.nameLocation = "end", axis.nameColor = NULL, line.width = 1,
  line.opacity = 0.8, axis.line = TRUE, axis.tick = FALSE,
  axis.lable = TRUE, axis.splitLine = TRUE, axis.splitArea = FALSE,
  axis.boundaryGap = TRUE, calculable = TRUE, showLabel = TRUE,
  opt = list())
}
\arguments{
\item{dat}{a data object (a matrix, a data frame or a factor array)}

\item{series}{an "formula" object: Associates the levels of variable
with symbol color, e.g. \code{series = ~groupName}}

\item{size}{an array of html widget width and height(either numeric pixels 
or percentage could be accepted): e.g. size = c(1024, 768).}

\item{theme}{an object of theme name. see(\url{http://datatables.net/extensions/index}) for detail.
supported theme: \code{c("default", "vintage", "dark", "westeros", "essos", "wonderland", "walden",
"chalk", "infographic", "macarons", "roma", "shine", "purple-passion")}}

\item{title}{an overall title for the plot. you can modify title widget after chart has been
generated: barEchart + eTitle(title = "your title.")}

\item{title.x, title.y}{the xy coordinates of main title, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify title widget after chart has been generated: 
\code{barEchart + eTitle(title="main title", x = "left", y=10)}}

\item{legend}{logical whether the legend widget show or not, default is TRUE.
you can modify legend widget after chart has been generated, the legend position and 
legend orientation are available at present.
\code{barEchart + eLegend(show = TRUE)}}

\item{legend.x, legend.y}{the xy coordinates of legend, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify legend widget after chart has been generated: 
\code{barEchart + eLegend(x = "right", y="top")}}

\item{legend.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify legend widget after chart has been generated: 
\code{barEchart + eLegend(orient = "vertical")}}

\item{calculable}{logical whether the front-end interactive chart will 
support the drag-recalculable feature.
the size and calculable option can be setted after bar chart has been 
generated through eOption: \code{barEchart + eOption(calculable = TRUE)}}

\item{xvar, yvar}{objects of class "formula" (or one that can be coerced 
to that class):  x,y coordinates of the given data.frame colnames, e.g. 
\code{xvar = ~xAxisName}; \code{yvar = ~yAxisName}. xvar, yvar only needed for the 
data.frame data input.}

\item{horiz}{logical. If FALSE, the bars are drawn vertically with the 
first bar to the left.}

\item{stackGroup}{list object, used to make series pre-stacked before rendering
whole bar chart, if \code{unique(df[["groupName"]])} = c("a", "b", "c", "d", 
"e", "f") and an example legal input for stackGroup should be: 
\code{stackGroup = list(c("a","b"), c("e","f"))}. And the c("a","b") and 
c("e","f") will be stacked into two bar, and the "c" and "d" will account 
two seperated bars.}

\item{toolbox}{logical whether the toolbox widget show or not, default is TRUE.
you can modify toolbox widget after chart has been generated, the toolbox position, toolbox 
element and toolbox orientation are available at present.  
\code{barEchart + eToolbox(show = TRUE)}}

\item{toolbox.x, toolbox.y}{the xy coordinates of toolbox, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify toolbox widget after chart has been generated: 
\code{barEchart + eToolbox(x = "right", y="top")}}

\item{toolbox.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify toolbox widget after chart has been generated: 
\code{barEchart + eToolbox(orient = "vertical")}}

\item{dataview, mark, restore, dataZoom, magicType}{logical variable whether the dataview
, mark, restore, dataZoom or magicType tool in toolbox widget show or not, 
default is TRUE. you can modify toolbox widget after chart has been generated, 
the toolbox position, toolbox element and toolbox orientation are available at present.  
\code{barEchart + eToolbox(dataView = FALSE)}}

\item{tooltip}{logical whether the tooltip widget for front-end interactive chart
show or not. default is TRUE. you can modify tooltip widget after chart has been generated, 
the tooltip trigger and tooltip formatter is available at present.  
\code{barEchart + eTooltip(show = FALSE)}}

\item{tooltip.trigger}{an element of c("axis", "item"), default is "axis" for bar chart.
"axis" option for trigger will show all the information of mouse;
"item" option for tirgger will only show the given item information of mouse.
you can modify tooltip widget after chart has been generated: 
\code{barEchart + eTooltip(trigger = "axis")}}

\item{tooltip.formatter}{the information formatter for tooltip widget, 
default is "<a>:<b><c>" for bar chart.
you can modify tooltip widget after chart has been generated: 
\code{barEchart + eTooltip(formatter = "<a><b>:<c>")}}
}
\description{
This function creates an HTML widget to display matrix, data.frame and 
factor array, using the JavaScript library ECharts3.
}
\note{
You are recommended to use lazyPlot function for interactive chart
  option set through "shiny" server.
}
\examples{
require(plyr)

axisList = list(
	list(index=7, type="category", data = c("low", "middle", "high")), 
	list(index=6, inverse=TRUE, max=50, nameLocation="start")
)
eParallel(head(parallelDf, 20), series=~groupName, axisList = axisList)

}
