% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePlot.R
\name{ePlot}
\alias{ePlot}
\title{Create rEcharts from an R object}
\usage{
ePlot (series, ext = NULL, size = NULL)
}
\arguments{
\item{series}{a list object for data in echarts}

\item{ext}{xAxis, yAxis, tooltip, toolbox, timeline setting in echarts}

\item{size}{an array of html widget width and height(either numeric pixels 
or percentage could be accepted): e.g. size = c(1024, 768).}
}
\examples{
series = list(
list(
  name = "Funnel",
  type = "funnel",
  size= c('80\%', '80\%'),
  textRotation = c(0, 45, 90, -45),
 textPadding = 0,
  data = list(list(value=100,
                   name="A"),
              list(value=66,
                   name="B"),
              list(value=33,
                   name="C"))
)
)
# ePlot(series)
# 
series2 = list(
  list(
    name = '\\u84b8\\u53d1\\u91cf',
    type = 'bar',
    data = c(2.0, 4.9, 7.0, 23.2, 25.6, 76.7, 135.6, 162.2, 32.6, 20.0, 6.4, 3.3)
  ),
  list(
    name = '\\u964d\\u6c34\\u91cf',
    type = 'bar',
    data = c(2.6, 5.9, 9.0, 26.4, 28.7, 70.7, 175.6, 182.2, 48.7, 18.8, 6.0, 2.3)

  ),
  list(
    name = '\\u5e73\\u5747\\u6e29\\u5ea6',
    type = 'line',
    yAxisIndex = 1,
    data = c(2.0, 2.2, 3.3, 4.5, 6.3, 10.2, 20.3, 23.4, 23.0, 16.5, 12.0, 6.2)

  )
)

ext = list(
  xAxis = list(list(
    type = "category",
    data = c('1\\u6708','2\\u6708','3\\u6708','4\\u6708','5\\u6708','6\\u6708','7\\u6708','8\\u6708','9\\u6708','10\\u6708','11\\u6708','12\\u6708')
  )),
  yAxis = list(
    list(
      type =  'value',
      name =  '\\u6c34\\u91cf',
      min =  0,
      max =  250,
      interval =  50,
      axisLabel =  list(
        formatter =  '{value} ml'
      )
    ),
   list(
      type =  'value',
      name =  '\\u6e29\\u5ea6',
      min =  0,
      max =  25,
      interval =  5,
      axisLabel = list(
        formatter= '{value} ℃'
      )
    )
  ))
ePlot(series2, ext )
}
