% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRadar.R
\name{eRadar}
\alias{eRadar}
\title{Radar charts}
\usage{
eRadar(dat, xvar = NULL, yvar = NULL, series = NULL, size = NULL,
  ymin = vector(), ymax = vector(), theme = "default", title = NULL,
  subtitle = NULL, title.x = "center", title.y = "top", legend = TRUE,
  legend.data = NULL, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  dataView = TRUE, readOnly = FALSE, mark = TRUE, dataZoom = FALSE,
  magicType = FALSE, tooltip = TRUE, tooltip.trigger = "item",
  formatter = "", axis.scale = TRUE, xlab = FALSE, ylab = FALSE,
  calculable = TRUE, showLabel = TRUE, opt = list())
}
\arguments{
\item{dat}{data.frame or matrix, should have colnames and rownames.}

\item{opt}{option of ECharts.}

\item{limit}{data.frame or matrix, 2 column, indicates the limit of each axis.}
}
\value{
The HTML code as a character string.
}
\description{
ECharts style radar charts.
}
\examples{
require(plyr)
dat = ddply(iris, .(Species), colwise(mean))
rownames(dat) = dat[,1]
dat = dat[, -1]
dat
eRadar(dat)
df2 <- data.frame(
 saleNum=c(10,20,30,40,50,60,70,15,25,35,45,55,65,75,25,35,45,55,65,75,85),
	seller=c(rep("Â∞èÈªÑËäë˘<U+3E31>",7), rep("Â∞èÁ∫¢",7), rep("Â∞èÁôΩ",7)),
weekDay = c(rep(c("Âë®‰∏Ä","Âë®‰∫å","Âë®‰∏â","Âë®Âõõ","Âë®‰∫î","Âë®ÂÖ≠","Âë®Êó•"),3))
)
dat <- df2
xvar=~weekDay; yvar= ~saleNum; series=~seller
eRadar(df2, ~weekDay, ~saleNum, ~seller)
}

