% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eSankey.R
\name{eSankey}
\alias{eSankey}
\title{Create an HTML sankey charts widget using the ECharts(version:3.2.2) library}
\usage{
eSankey(dat, candleColor, size = NULL, title = NULL, subtitle = NULL,
  title.x = "center", title.y = "top", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  tooltip = TRUE, tooltip.trigger = "item", calculable = TRUE,
  opt = list(), ...)
}
\arguments{
\item{dat}{a data.frame object}

\item{size}{an array of html widget width and height(either numeric pixels 
or percentage could be accepted): e.g. size = c(1024, 768).}

\item{title}{an overall title for the plot. you can modify title widget after chart has been
generated: sankeyEchart + eTitle(title = "your title.")}

\item{title.x, title.y}{the xy coordinates of main title, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify title widget after chart has been generated: 
\code{sankeyEchart + eTitle(title="main title", x = "left", y=10)}}

\item{toolbox}{logical whether the toolbox widget show or not, default is TRUE.
you can modify toolbox widget after chart has been generated, the toolbox position, toolbox 
element and toolbox orientation are available at present.  
\code{sankeyEchart + eToolbox(show = TRUE)}}

\item{toolbox.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify toolbox widget after chart has been generated: 
\code{sankeyEchart + eToolbox(orient = "vertical")}}

\item{toolbox.x, toolbox.y}{the xy coordinates of toolbox, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify toolbox widget after chart has been generated: 
\code{sankeyEchart + eToolbox(x = "right", y="top")}}

\item{tooltip}{logical whether the tooltip widget for front-end interactive chart
show or not. default is TRUE. you can modify tooltip widget after chart has been generated, 
the tooltip trigger and tooltip formatter is available at present.  
\code{sankeyEchart + eTooltip(show = FALSE)}}

\item{tooltip.trigger}{an element of c("axis", "item"), default is "item" for sankey chart.
"axis" option for trigger will show all the information of mouse;
"item" option for tirgger will only show the given item information of mouse.
you can modify tooltip widget after chart has been generated: 
\code{sankeyEchart + eTooltip(trigger = "item")}}

\item{calculable}{logical whether the front-end interactive chart will 
support the drag-recalculable feature.
the size and calculable option can be setted after bar chart has been 
generated through eOption: \code{sankeyEchart + eOption(calculable = TRUE)}}
}
\description{
This function creates an HTML widget to display data.frame, using the JavaScript library ECharts3.
}
\note{
You are recommended to use lazyPlot function for interactive chart
  option set through "shiny" server.
}
\examples{
dat = data.frame(source=c("Agricultural 'waste'","Bio-conversion",
                          "Bio-conversion","Bio-conversion","Bio-conversion",
                          "Biofuel imports","Biomass imports","Coal imports",
                          "Coal reserves","Coal","District heating","District heating",
                          "District heating","Electricity grid","Electricity grid"),
                 target=c("Bio-conversion","Liquid","Losses","Solid","Gas","Liquid",
                          "Solid","Coal","Coal","Solid","Industry","Heating and cooling - commercial",
                          "Heating and cooling - homes","Over generation / exports","Heating and cooling - homes"),
                 value=c(124.729,0.597,26.862,280.322,81.144,35,35,11.606,63.965,75.571,
                         10.639,22.505,46.184,104.453,113.726))
eSankey(dat)
}
