% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charts.R
\name{eScatter}
\alias{eScatter}
\title{Create an HTML line charts widget using the ECharts(version:3.2.2) library}
\usage{
eScatter(...)
}
\arguments{
\item{dat}{a data object (a matrix, a data frame or a factor array)}

\item{xvar, yvar}{objects of class "formula" (or one that can be coerced
to that class):  x,y coordinates of the given data.frame colnames, e.g.
\code{xvar = ~xAxisName}; \code{yvar = ~yAxisName}. xvar, yvar only needed for the
data.frame data input.}

\item{series}{an "formula" object: Associates the levels of variable
with symbol color, e.g. \code{series = ~groupName}}

\item{size}{an array of html widget width and height(either numeric pixels
or percentage could be accepted): e.g. size = c(1024, 768).}

\item{theme}{an object of theme name. see(\url{http://datatables.net/extensions/index}) for detail.
supported theme: \code{c("default", "vintage", "dark", "westeros", "essos", "wonderland", "walden",
"chalk", "infographic", "macarons", "roma", "shine", "purple-passion")}}

\item{title}{an overall title for the plot. you can modify title widget after chart has been
generated: Echart + eTitle(title = "your title.")}

\item{title.x, title.y}{the xy coordinates of main title, besides the excat exact pixels value,
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input.
you can modify title widget after chart has been generated:
\code{Echart + eTitle(title="main title", x = "left", y=10)}}

\item{legend}{logical whether the legend widget show or not, default is TRUE.
you can modify legend widget after chart has been generated, the legend position and
legend orientation are available at present.
\code{Echart + eLegend(show = TRUE)}}

\item{legend.x, legend.y}{the xy coordinates of legend, besides the excat exact pixels value,
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input.
you can modify legend widget after chart has been generated:
\code{Echart + eLegend(x = "right", y="top")}}

\item{legend.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify legend widget after chart has been generated:
\code{Echart + eLegend(orient = "vertical")}}

\item{toolbox}{logical whether the toolbox widget show or not, default is TRUE.
you can modify toolbox widget after cahart has been generated, the toolbox position, toolbox
element and toolbox orientation are available at present.
\code{Echart + eToolbox(show = TRUE)}}

\item{toolbox.x, toolbox.y}{the xy coordinates of toolbox, besides the excat exact pixels value,
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input.
you can modify toolbox widget after chart has been generated:
\code{Echart + eToolbox(x = "right", y="top")}}

\item{toolbox.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify toolbox widget after chart has been generated:
\code{Echart + eToolbox(orient = "vertical")}}

\item{dataview, mark, restore, dataZoom, magicType}{logical variable whether the dataview
, mark, restore, dataZoom or magicType tool in toolbox widget show or not,
default is TRUE. you can modify toolbox widget after chart has been generated,
the toolbox position, toolbox element and toolbox orientation are available at present.
\code{Echart + eToolbox(dataView = FALSE)}}

\item{tooltip}{logical whether the tooltip widget for front-end interactive chart
show or not. default is TRUE. you can modify tooltip widget after chart has been generated,
the tooltip trigger and tooltip formatter is available at present.
\code{Echart + eTooltip(show = FALSE)}}

\item{tooltip.trigger}{an element of c("axis", "item"), default is "axis" for line chart.
"axis" option for trigger will show all the information of mouse;
"item" option for tirgger will only show the given item information of mouse.
you can modify tooltip widget after chart has been generated:
\code{Echart + eTooltip(trigger = "axis")}}

\item{tooltip.formatter}{the information formatter for tooltip widget,
default is "<a>:<b><c>" for line chart.
you can modify tooltip widget after chart has been generated:
\code{Echart + eTooltip(formatter = "<a><b>:<c>")}}

\item{calculable}{logical whether the front-end interactive chart will
  support the drag-recalculable feature.
  the size
Create the scatter chart}
}
\description{
This function creates an HTML widget to display matrix, data.frame and
factor array, using the JavaScript library ECharts3.
}

