% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{events}
\alias{events}
\title{Events Data for eventRiver}
\format{A data.frame (30 * 5) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 series \tab chr \tab Type of the events ("financial", "political") \cr
 event \tab chr \tab Title of the events ("Alibaba IPO", ...) \cr
 weight \tab num \tab The weight of the events \cr
 time \tab Date \tab Time of the event slices \cr
 value \tab int \tab Impact value of the event slices
}}
\description{
A data.frame comprising of fictious events.
}
\examples{
data(events)
str(events)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_05_eventRiver.html}
}

