% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flight}
\alias{flight}
\title{Fictious Flights in China (Beijing, Shanghai, Guangzhou)}
\format{A data.frame (150 * 2) route \cr
\tabular{ll}{
 Var \tab Type \tab Meaning \cr
 From \tab chr \tab Source of the route (Beijing, Shanghai, ...) \cr
 To \tab chr \tab Target of the route (Baotou, Beihai, ...)
}
A data.frame (114 * 3) coord \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 Place \tab chr \tab The name of the cities (also in route$From and route$To) \cr
 Lng \tab num \tab Longitude of the places \cr
 Lat \tab num \tab Latitude of the places
}}
\description{
A list comprising of two data.frame: route among cities and coordinates of the cities.
}
\examples{
data(flight)
str(flight)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_31_Map.html}
}

