% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geoNameMap}
\alias{geoNameMap}
\title{Chinese and English Names of Geographic Names}
\format{A data.frame (3735 * 5) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 ID \tab int \tab ID of the record \cr
 EN \tab chr \tab English geographic names \cr
 CN \tab chr \tab Chinese translation of the geographic names \cr
 FKEY \tab int \tab ID of the parent node of this place \cr
 LEVEL \tab int \tab 0: country; 1: province; 2: city; 3: county
}}
\description{
A data.frame acting as a dictionary of Chinese-English-names of a bunch of geograhic names.
}
\examples{
data(geoNameMap)
str(geoNameMap)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_31_Map.html}
}

