% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{\%+=\%}
\alias{\%**=\%}
\alias{\%*=\%}
\alias{\%+=\%}
\alias{\%-=\%}
\alias{\%/=\%}
\alias{\%\%=\%}
\alias{\%\=\%}
\alias{\%^=\%}
\alias{\%mod=\%}
\alias{\%root=\%}
\title{Javascript Operator Assignment}
\usage{
lhs \%+=\% rhs

lhs \%-=\% rhs

lhs \%*=\% rhs

lhs \%/=\% rhs

lhs \%^=\% rhs

lhs \%**=\% rhs

lhs \%\\=\% rhs

lhs \%mod=\% rhs

lhs \%root=\% rhs

lhs \%\%=\% rhs
}
\arguments{
\item{lhs}{Left hand side argument.}

\item{rhs}{Right hand side argument.}

\item{envir}{Environment for operation to take place.}
}
\value{
Modified lhs
}
\description{
If you are familiar with \code{Javascript}, you may miss +=, -=, *=, /= very much.
The operator can be \code{+, -, *, /, ^, **, \\, mod, root}.
}
\examples{
\dontrun{
a <- 3
a \%+=\% 2     # 3 > (3 + 2) > 5
a \%-=\% 2     # 5 > (5 - 2) > 3
a \%*=\% 2     # 3 > (3 * 2) > 6
a \%/=\% 2     # 6 > (6 / 2) > 3
a \%^=\% 2     # 3 > (3 ^ 2) > 9
a \%**=\% 2    # 9 > (3 ** 2) > 81
a \%\\=\% 30    # 81 > (81 \%\% 30) > 21
a \%mod=\% 8   # 21 > (21 \%\% 8) > 5
a \%root=\% 2  # 5 > (5 ^ (1/2)) > 2.236

}
}

