% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lazyPlot.R
\name{lazyPlot}
\alias{lazyPlot}
\title{Modify the echarts object through shiny server}
\usage{
lazyPlot(e1)
}
\arguments{
\item{e1}{a recharts output object, could be generated by eBar, eLine, eRadar, eMap, eArea, ePoints,
eFunnel, eForce and other echarts generate functions(will be added later).}
}
\description{
Launch a shiny server to modify the generated echart object from the front-end interface,
which integrate the general chart optional settings, also the code will show on the page.
}
\note{
You are recommended to use lazyPlot function for interactive chart
  option set through "shiny" server.
}
\examples{
  mapData <- head(mapTestData_chs, 5)
  echartObj <- eMap(mapData, namevar=~stdName, datavar = ~val1 + val2)
  lazyPlot(echartObj) # will launch the shiny server to
}
