% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mapTestData_chs}
\alias{mapTestData_chs}
\title{mapTestData_chs use for eMap chart generation.}
\format{Data frame with 3380 observations of 5 variables.}
\source{
\url{http://www.justice.gov/dea/resource-center/stride-data.shtml}
}
\usage{
mapTestData_chs
}
\description{
This dataset comes from geo infomation.
}
\keyword{datasets}
