% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mideast}
\alias{mideast}
\title{For- and Against- Relationship in Mid-east (2013)}
\format{A matrix (16 * 16) \cr
Mutual attitude of the political powers (for or against/weight)}
\description{
A matrix of the relationship (for- or against-) of the political powers involved
in Mid-east. You have to re-organize it before processing.
}
\examples{
data(mideast)
str(mideast)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_12_Chord.html}
}

