% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartOpt.R
\name{optionSet}
\alias{optionSet}
\title{recharts set fucntion}
\usage{
optionSet(chart, ...)
}
\arguments{
\item{obj}{recharts object.}
}
\value{
The output list of recharts as a list.
}
\description{
An shell function for setting arguments of the recharts object.
}
