% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parallelDf}
\alias{parallelDf}
\title{parallelDf use for eParallel chart generation.}
\format{Data frame with 1140 observations of 6 variables.}
\usage{
parallelDf
}
\description{
This dataset comes from  infomation.
}
\keyword{datasets}
