\name{plot.recharts}
\alias{plot.recharts}
\title{recharts plot fucntion}
\usage{
  plot.recharts(x, tag = NULL, ...)
}
\arguments{
  \item{x}{recharts plot object.}

  \item{tag}{whether plot the recharts object to browser or
  string.}

  \item{...}{default parameter.}
}
\value{
  The HTML code as a character string or an html page.
}
\description{
  An shell function for ploting of the recharts object.
}

