% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.recharts.R
\name{+.echarts}
\alias{+.echarts}
\title{Modify a recharts by adding on new components.}
\usage{
\method{+}{echarts}(e1, e2)
}
\arguments{
\item{e1}{An object of class \code{recharts}}

\item{e2}{A component to add to \code{e1}}
}
\description{
Modify a recharts by adding on new components.
}
\seealso{
\code{\link{set}}
}
