\name{print.recharts}
\alias{print.recharts}
\title{recharts print fucntion}
\usage{
  print.recharts(x, tag = NULL, file = "", ...)
}
\arguments{
  \item{x}{recharts print object.}

  \item{tag}{whether print the recharts object to browser
  or string.}

  \item{...}{default parameter.}
}
\value{
  The HTML code as a character string.
}
\description{
  An shell function for printing of the recharts object.
}

