% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{recharts}
\alias{recharts}
\alias{recharts-package}
\title{recharts: An R Interface to ECharts 2}
\description{
\pkg{recharts} provides an interface to the JavaScript library ECharts 2
(developed by Baidu Inc.) for data visualization.
}
\details{
The goal of this package is to make it easy to create charts with only
a few lines of R code. R users should be able to get started with ECharts
without having to know HTML or JavaScript, although advanced users will benefit
from their knowledge of JavaScript.
\cr \cr
The package \pkg{recharts} was invented by Taiyun Wei and Yang Zhou. And this
version was forked from Yihui Xie's repo.
\cr \cr
You can find the manual by \code{browseVignettes("recharts")}.
}
\references{
You can also visit the official dev site
\url{http://madlogos.github.io/recharts}(Eng) /
\url{http://madlogos.github.io/recharts/index_cn.html}(Chs) to read the manual.
}
\author{
\strong{Maintainer} Yiying Wang \email{madlogos@gmail.com}
}
