% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveREcharts.R
\name{saveREcharts}
\alias{saveREcharts}
\title{Save recharts into a png/jpg file}
\usage{
saveREcharts(chart, file, size = NULL, delay = 2)
}
\arguments{
\item{chart}{a recharts object}

\item{file}{file name for saving}

\item{size}{htmlwidgets size for recharts}

\item{delay}{Time to wait before taking screenshot, in seconds. Sometimes a longer delay is needed for all assets to display properly.}
}
\description{
This function creates an HTML widget to display matrix, data.frame and 
factor array, using the JavaScript library ECharts3.
}
