% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{setColors}
\alias{setColors}
\title{Set Colors for Echarts}
\usage{
setColors(palette)
}
\arguments{
\item{palette}{Palette, default NULL. Could be
\itemize{
 \item palette name, e.g, "Blues"
 \item a hex color, e.g., "#FFFFFF", "0xFFFFFFFF"
 \item a vector or color names, hex colors
 \item NULL
}}
}
\value{
A vector of hex colors
}
\description{
Set color vector from a palette.
}
\examples{
\dontrun{
setColors(NULL)
setColors("terrain")
setColors(c('red', 'yellow', 'blue'))
}
}
\seealso{
\code{\link{recharts:::setPal}} \code{\link{RColorBrewer}} \code{\link{ggthemes}}
}

