% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.R
\name{setOptions}
\alias{setOptions}
\title{set options for echart}
\usage{
setOptions(chart, x)
}
\description{
List or JsonString
optionType = c("title","legend","grid","xAxis","yAxis","polar","radiusAxis","angleAxis","radar",
         "dataZoom", "visualMap", "toolbox","tooltip","brush","geo", "parallel", "parallelAxis",
         "singleAxis","timeline","textStyle","series","color","backgroundColor",
         "animation","animationThreshold","animationDuration","animationEasing","animationDelay",
         "animationDurationUpdate","animationEasingUpdate","animationDelayUpdate",
         "progressive","progressiveThreshold","blendMode","hoverLayerThreshold")
}

