% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{setPal}
\alias{setPal}
\title{Set Palettes (not export)}
\usage{
setPal(palname, n = 6)
}
\arguments{
\item{palname}{name of the palette}

\item{n}{length of the color vector, default 6}
}
\value{
color vectors
}
\description{
Set Palettes (not export)
}
\examples{
\dontrun{
setPal('tableau20')
}
}
\seealso{
\code{\link{RColorBrewer}} \code{\link{scales}} \code{\link{ggthemes}}
}

