% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTheme}
\alias{setTheme}
\title{Set Theme of Echarts}
\usage{
setTheme(chart, theme = c("default", "macarons", "infographic", "blue",
  "dark", "gray", "green", "helianthus", "macarons2", "mint", "red", "roma",
  "sakura", "shine"), ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{theme}{\code{'default', 'macarons', 'infographic', 'blue', 'dark', 'gray',
'green', 'helianthus', 'macarons2', 'mint', 'red', 'roma', 'sakura', 'shine'}}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set the theme of Echarts. \cr
When an echart object is generated, you can modify it by setting toolbox using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
g <- echartR(iris, Sepal.Width, Petal.Width)
g \%>\% setTheme("infographic")
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/example/theme.html}
}

