% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTimeline}
\alias{setTimeline}
\title{Set \code{timeline} of Echarts}
\usage{
setTimeline(chart, show = TRUE, type = c("time", "number"),
  realtime = TRUE, grid = list(80, NULL, 80, 0), width = NULL,
  height = 50, bgColor = "rgba(0,0,0,0)", borderColor = "#ccc",
  borderWidth = 0, controlPosition = c("left", "right", "none"),
  autoPlay = FALSE, loop = TRUE, playInterval = 2000, lineStyle = NULL,
  label = NULL, checkpointStyle = NULL, controlStyle = NULL,
  symbol = "emptyDiamond", symbolSize = 4, currentIndex = 0,
  data = NULL, ...)
}
\arguments{
\item{show}{Logical. If or not the timeline is shown. Default TRUE.}

\item{type}{'time' or 'number' format of the timeline. Default 'time'.}

\item{realtime}{Logical. If or not the changes take effect in realtime manner.
Default TRUE.}

\item{grid}{A list of \code{x, y, x2, y2}. x, y refers to the coorindate of the
upper-left point while x2, y2 refers to that of the lower-right point. Default
\code{list(80, NULL, 80, 0)}.}

\item{width}{Width of the timeline. Could be a number or a character in percent
form. Default NULL.}

\item{height}{Height of the timeline. Default 50 (px).}

\item{bgColor}{Background color of the timeline. Default "rgba(0,0,0,0)" (transparent).}

\item{borderColor}{Border color of the timeline. Default "#ccc".}

\item{borderWidth}{Border width of the timeline. Default 0 (px) (not shown).}

\item{controlPosition}{Position of the control of the timeline. Could be 'left',
'right' or 'none'. Default 'left'.}

\item{autoPlay}{Logical. If or not the timeline auto displays. Default FALSE.}

\item{loop}{Logical. If or not the timeline displays in loop mode. Default TRUE.}

\item{playInterval}{Interval when displays each timeslice. Default 2000 (ms).}

\item{lineStyle}{A list. Line style of the timeline. Default value: \cr
\code{list(color="#666", width=1, type="dashed")}. \cr Supports features of \code{'color',
'width', 'type', 'shadowColor', 'shadowBlur', 'shadowOffsetX', 'shadowOffsetY'}}

\item{label}{A list. Label style of the timeline. Default value: \cr
\code{list(show=true, interval="auto", rotate=0, formatter=NULL, \cr
textStyle=list(color="#333"))}. \cr Supports features of \code{'show', 'interval',
 'rotate', 'formatter', 'textStyle'}.}

\item{checkpointStyle}{A list. Checkpoint style of the timeline. Default value: \cr
\code{list(symbol="auto", symbolSize="auto", color="auto", \cr
borderColor="auto", borderWidth="auto", \cr
label=list(show=FALSE, textStyle=list(color="auto")))}. \cr
Supports features of \code{'symbol', 'symbolSize', 'color', 'borderColor',
'borderWidth', 'label'}.}

\item{controlStyle}{A list. Control style of the timeline. Default value: \cr
\code{list(itemSize=15, itemGap=5, normal=list(color="#333"),
emphasis=list(color="#1e90ff"))}. \cr
Supports features of \code{'itemSize', 'itemGap', 'normal', 'emphasis'}.}

\item{symbol}{Character. The symbol used in timeline. Default 'emptyDiamond'.
You can use symbols in \code{\link{setSymbolList}}.}

\item{symbolSize}{The size of the symbols. Default 4.}

\item{currentIndex}{The current index position, in correspondance with \code{z}.
It is used to show specific timeline slices. Default 0.}

\item{data}{The data list of the timeline, also used as timeline data label.
Default NULL.}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set timeline of Echarts when the echarts object contains timeline slices (z). \cr
When an echart object is generated, you can modify it by setting tooltip using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
iris$tag <- 1 + as.integer(row.names(iris)) \%\% 3
iris <- iris[order(iris$tag),]
g <- echartR(iris, Sepal.Width, Petal.Width, series=Species, z=tag)
g \%>\% setTimeline(symbol='emptyCircle')
}

}

