% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{setTitle}
\alias{setTitle}
\title{Set Title And Subtitle of Echarts}
\usage{
setTitle(chart, title = NULL, subtitle = NULL, link = NULL,
  sublink = NULL, pos = 6, ...)
}
\arguments{
\item{chart}{\code{echart} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{title}{text of the title. If written in markdown format \code{[caption](url)},
then \code{caption} is passed to title, \code{url} is passed to link. If the length
of the title vector equals to the length of timeline slices, the title vector will
be used as slice-specific user-defined title.}

\item{subtitle}{text of the subtitle. If written in markdown format \code{[caption](url)},
then \code{caption} is passed to title, \code{url} is passed to link. If the length
of the subtitle vector equals to the length of timeline slices, the subtitle vector will
be used as slice-specific user-defined title.}

\item{link}{link of the title}

\item{sublink}{link of the subtitle}

\item{pos}{the clock-position of title (and subtitle), refert to \code{\link{vecPos}}}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting title and
subtitles using \%>\%.
}
\examples{
\dontrun{
## simple echarts scatter plot
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')

## set simple titles
g \%>\% setTitle(title='Iris data set', subtitle='by: R. A. Fisher')

## set titles with links
g \%>\% setTitle(
  '[Iris data set](https://en.wikipedia.org/wiki/Iris_flower_data_set)',
  '[R. A. Fisher](https://en.wikipedia.org/wiki/Ronald_Fisher)')

## echarts with timeline
g1 <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, z=Species, type='scatter')

## simple titles/subtitles
g1 \%>\% setTitle(
    '[Iris data set](https://en.wikipedia.org/wiki/Iris_flower_data_set)',
    '[R. A. Fisher](https://en.wikipedia.org/wiki/Ronald_Fisher)')

## user-defined titles/subtitles
titles <- c(
    '[Iris species 1](https://en.wikipedia.org/wiki/Iris_setosa)',
    '[Iris species versicolor](https://en.wikipedia.org/wiki/Iris_versicolor)',
    '[Iris species 3](https://en.wikipedia.org/wiki/Iris_virginica)')
g1 \%>\% setTitle(titles)
}
}

