% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setToolbox}
\alias{setToolbox}
\title{Set \code{toolbox} of Echarts}
\usage{
setToolbox(chart, show = TRUE, language = "cn", controls = c("mark",
  "dataZoom", "dataView", "magicType", "restore", "saveAsImage"), pos = 1,
  ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{logical. Show the toolbox if TRUE.}

\item{language}{'cn' or 'en', the language of the toolbox tooltips.}

\item{controls}{which widgets to show. Default \code{'mark', 'dataZoom', 'dataView', 'magicType',
'restore', 'saveAsImage'}.}

\item{pos}{the clock-position of toolbox, refer to \code{\link{vecPos}}. Or you can
define a vector \code{c(x, y, orient)} yourself.}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting toolbox using
\code{\link{\%>\%}}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setToolbox(TRUE, 'en')
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~toolbox}
}

