% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock}
\alias{stock}
\title{[Dataset] Stock Index Records of Shanghai Stock Exchange (2013H1)}
\format{A data.frame (88 * 5) \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 data \tab Data \tab Transaction date \cr
 open \tab num \tab Open index \cr
 close \tab num \tab Close index \cr
 low \tab numt \tab Low index \cr
 high \tab num \tab High index
}}
\description{
A data.frame of stock index records of SHSE in 2013H1.
}
\examples{
data(stock)
str(stock)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_04_K.html}
}
