% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSetUtils.R
\name{tilteSet}
\alias{tilteSet}
\title{recharts tilteSet fucntion}
\usage{
tilteSet(title = NULL, link = NULL, subtitle = NULL, sublink = NULL,
  title.x = "left", title.y = "top")
}
\arguments{
\item{title}{The main title text.}

\item{link}{The hyper link of main title text.}

\item{subtitle}{Subtitle text.}

\item{sublink}{The hyper link of subtitle text.}

\item{title.x}{to be confirmed.}

\item{title.y}{to be confirmed.}
}
\value{
The list for tilteSet.
}
\description{
An shell function for printing of the recharts object.
}
