% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{title}
\alias{title}
\title{Create a title/subtitle for a chart}
\usage{
title(chart, show = TRUE, text = "", link = "", subtext = "",
  sublink = "", x = c("center", "left", "right"), y = c("top", "bottom",
  "center"), textStyle.fontSize = 12, textStyle.fontStyle = c("normal",
  "italic", "oblique"), textStyle.fontWeight = c("normal", "bold", "bolder",
  "lighter"))
}
\arguments{
\item{text}{Title}

\item{link}{Title hyperlink}

\item{subtext}{Subtitle}

\item{sublink}{Subtitle hyperlink}
}
\description{
Add a title/subtitle to a chart,
support Chinese character under linux, don't sure about windows.
}

