% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/options.R
\name{tooltip}
\alias{tooltip}
\title{Create a tooltip for a chart}
\usage{
tooltip(chart, show = TRUE, showContent = TRUE, trigger = c("item",
  "axis"), position = NULL, showDelay = 20, hideDelay = 100,
  transitionDuration = 0.4, enterable = FALSE,
  backgroundColor = "rgba(0,0,0,0,7)", borderColor = "#333",
  borderRadius = 4, borderWidth = 0, padding = 5,
  textStyleColor = "#fff")
}
\arguments{
\item{trigger}{item(default) or axis}
}
\description{
Add the tooltip,(mouse hover show the details)
}

