% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{tuneGrid}
\alias{tuneGrid}
\alias{tune_grid}
\title{Tune The Position of An Echarts Object}
\usage{
tuneGrid(chart, ...)

tune_grid(chart, ...)
}
\arguments{
\item{chart}{An Echarts object}
}
\description{
The funcion is used to tune x, y, x2, x2 of an Echarts object regarding its controls,
such as title, timeline, legend, toolbox, dataRange, dataZoom, roamController.
}
