% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vecPos}
\alias{vecPos}
\title{Text Position and Direction}
\usage{
vecPos(pos)
}
\arguments{
\item{pos}{1-12, clock digits.}
}
\value{
A vector of x-alignment, y-alignment and direction.
}
\description{
Converts text postion from clock digits to c(x, y, direction) vector
}
\note{
# Postion of Clock Numbers 1-12 \cr
\tabular{lllll}{
 10(l, t, v) \tab 11(l, t, h) \tab 12(c, t, h) \tab 1(r, t, h) \tab 2(r, t, v) \cr
 9(l, c, v) \tab \tab \tab \tab 3(r, c, v) \cr
 8(l, b, v) \tab 7(l, b, h) \tab 6(c, b, h) \tab 5(r, b, h) \tab 4(r, b, v)
}
}
\examples{
\dontrun{
vecPos(2) ## returns c("right", "top", "vertical")
}
}

