% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordFreqDf}
\alias{wordFreqDf}
\title{wordFreqDf for eWordcloud chart generation.}
\format{Data frame with 337 observations of 2 variables.}
\usage{
wordFreqDf
}
\description{
This dataset comes from ECharts original data.
}
\keyword{datasets}
