% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wordFreqDf_chs}
\alias{wordFreqDf_chs}
\title{wordFreqDf_chs for eWordcloud chart generation.}
\format{Data frame with 996 observations of 2 variables.}
\usage{
wordFreqDf_chs
}
\description{
This dataset comes from published articles of cos.name.
And use the Chinese word segmentation package(jiebaR) to generate the word data.frame.
}
\keyword{datasets}
