% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yuNetwork}
\alias{yuNetwork}
\title{Family Network of the Yus of Shaoxing}
\format{A data.frame (49 * 3) nodes \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 name \tab chr \tab The name of the family member/relative \cr
 series \tab chr \tab Type of the person ('root', 'node 1', ...) \cr
 value \tab num \tab Importance value of the person
}
A data.frame (49 * 4) links \cr
\tabular{lll}{
 Var \tab Type \tab Meaning \cr
 source \tab chr \tab Starting end of the link between two people \cr
 target \tab chr \tab Ending end of the link between two people \cr
 relation \tab chr \tab Relationship (father-son, spouse, ...) \cr
 weight \tab num \tab Importance of the relationship
}}
\description{
A list comprising of two data.frame: nodes (people related to the family) and
links (their relationship). The Yus have contributed significantly to connecting
a lot of most important people in modern China history.
}
\examples{
data(yuNetwork)
str(yuNetwork)
}
\references{
\url{https://madlogos.github.io/recharts/Basic_Plots_11_Force.html}
}

