% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLine.R
\name{eArea}
\alias{eArea}
\title{Area charts}
\usage{
eArea(dat, xvar = NULL, yvar = NULL, series = NULL, size = NULL,
  horiz = FALSE, stack = "SUM", theme = "default", title = NULL,
  subtitle = NULL, title.x = "center", title.y = "top", legend = TRUE,
  legend.x = "left", legend.y = "top", legend.orient = "horizontal",
  toolbox = TRUE, toolbox.orient = "horizontal", toolbox.x = "right",
  toolbox.y = "top", dataView = TRUE, readOnly = FALSE, mark = TRUE,
  dataZoom = FALSE, magicType = TRUE, tooltip = TRUE,
  tooltip.trigger = "item", formatter = "", axis.scale = FALSE,
  axis.line = TRUE, axis.tick = FALSE, axis.lable = TRUE,
  axis.splitLine = TRUE, axis.splitArea = FALSE, axis.boundaryGap = TRUE,
  xlab = TRUE, xlab.type = "category", xlab.data = NULL,
  xlab.position = "bottom", xlab.name = "", xlab.namePosition = "start",
  xlim = NULL, ylab = TRUE, ylab.type = "value", ylab.data = NULL,
  ylab.position = "left", ylab.name = "", ylab.namePosition = "start",
  ylim = NULL, calculable = TRUE, showLabel = TRUE, opt = list())
}
\arguments{
\item{dat}{data.frame or matrix}

\item{opt}{option of ECharts.}
}
\value{
The HTML code as a character string.
}
\description{
ECharts style area charts.
}
\examples{
eArea(WorldPhones)
}

