% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{echartR}
\alias{echartR}
\title{A lattice-like workaround function to generate Echarts}
\usage{
echartR(data, x = NULL, y = x, z = NULL, series = NULL, weight = NULL,
  lat = NULL, lng = NULL, x1 = NULL, lat1 = NULL, lng1 = NULL,
  type = "auto", stack = FALSE, xlab = NULL, xAxis = list(lab = xlab,
  color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0), ylab = NULL,
  yAxis = list(lab = ylab, color = NULL, splitLine = TRUE, banded = FALSE,
  rotate = 0), xlab1 = NULL, xAxis1 = list(lab = xlab1, series = NULL,
  reverse = FALSE, color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0),
  ylab1 = NULL, yAxis1 = list(lab = ylab1, series = NULL, reverse = FALSE,
  color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0), xyflip = FALSE,
  AxisAtZero = c(FALSE, TRUE), scale = TRUE, tooltip = TRUE,
  pos = list(title = 6, legend = 11, toolbox = 1, dataZoom = 6, dataRange = 8,
  roam = 2), calculable = TRUE, asImage = FALSE,
  markLine = markLinesmooth, markLinesmooth = NULL, markPoint = NULL,
  theme = list(backgroundColor = NULL, borderColor = NULL, borderWidth = 1,
  width = NULL, height = NULL), ...)
}
\arguments{
\item{data}{data.frame. No defaults.}

\item{x}{x variable, only omitable for histograms, pie, ring and rose charts.}

\item{y}{y variable. If y is omitted, assign x to y.}

\item{z}{z variable, only accept data/time variable to open time axis. Optional.}

\item{series}{series variable for grouping.}

\item{weight}{weight variable, used in histogram, bubble, etc.}

\item{lat}{Lattitude variable, only for point/line-marking map.}

\item{lng}{Longitude variable, only for point/line-marking map.}

\item{x1}{Backup x variable, only for line-marking map, force andchord chart.}

\item{lat1}{Backup lattitude variable, only for line-marking map.}

\item{lng1}{Backup Longitude variable, only for line-marking map.}

\item{type}{options \code{'scatter', 'point', 'bubble', 'bar', 'line', 'map',
'k', 'pie', 'ring', 'rose','area', 'chord', 'force', 'tree', 'treemap',
'wordcloud', 'heatmap', 'histogram', 'funnel', 'pyramid', 'radar'} \cr
\itemize{
  \item If 'map' is chosen, the control option should be a vector of length 3:
  \code{c('map', mapType, markType)}. \code{mapType} could be either 'world' or 'china',
  of which simplified Chinese names are required for 'china'.
  When markType equals to 'area', the function colors polygons to show the effects;
  while equals to 'point', it ticks pins on the map; while equals to 'line',
  it ticks lines on the map.
  \item If 'line', 'area' are chosen, the control option should be a vector of
  length 2: \code{c(main type, type feature)}. \code{main type} is line or area,
  \code{type feature} is 'smooth'.
  \item If 'chord' is chosen, the control option should be a vector of
  length 2: \code{c('chord', 'riboon')}, which refers to chord chart with ribbon.
  \item If 'radar' is chosen, the control option should be a vector of
  length 2: \code{c('radar', 'fill')}, which refers to radar chart with color filled.
}}

\item{stack}{Default to FALSE (do not stack). Used in stacked column, bar, line
and area chart, etc.}

\item{xlab}{You can also omit xAxis, directly assign xAxis title.
xlab has a higher priority than xAxis[['lab']].}

\item{xAxis}{x Axis parameters in a list, default
\code{list(lab=NULL, color=NULL, splitLine=TRUE, banded=FALSE, rotate=0)}.}

\item{ylab}{You can also omit yAxis, directly assign yAxis title.
ylab has a higher priority than yAxis[['lab']].}

\item{yAxis}{parameters of y Axis. Refer to xAxis.}

\item{xlab1}{Title of secondary x-axis, with higher priority than lab in xAxis1 list,}

\item{xAxis1}{Secondary x axis, effecitve only if there are multiple series.
Written in a list, default \code{list(lab=NULL, series=NULL, reverse=FALSE,
color=NULL, splitLine=TRUE, banded=FALSE, rotate=0)}.
series set which series to put on secondary x-axis, which can be either a name
vector or an index vector, e.g., c(1,2) or c('male','female').
The data is fliped up when reverse is set TRUE.}

\item{ylab1}{Title of secondary y-axis, with higher priority than lab in yAxis1 list,}

\item{yAxis1}{Refer to xAxis1}

\item{xyflip}{Flip x, y-axies. Default FALSE.}

\item{AxisAtZero}{Axes (and sub axes) cross at zero. Default \code{c(FALSE, TRUE)}.}

\item{scale}{Rescale the axes based on min and max values. Default TRUE.}

\item{tooltip}{Mouse tip effects swtich. Default TRUE.}

\item{pos}{Position of image elements which are integers 1-12.
Title position default to 6 (o'clock), which means bottom middle.
Defaults to legend, toolbox, dataRange, dataZoom, roamController are
11, 1, 6, 8, 2, respectively. Default pos=list(title=6, legend=11, toolbox=1,
dataZoom=6, dataRange=8, roam=2)}

\item{calculable}{Calculable switch (Echarts patent).}

\item{asImage}{renderAsImage switch.Deafult to FALSE.}

\item{markLine}{Short form: \cr
\tabular{llll}{
 [col 1] series name|index \tab [col 2] line name \tab [col 3] Line type
 \tab [col 4] Light effect \cr
 String or number \tab String|NA \tab min|max|average|lm \tab TRUE|FALSE
}
Full form: \cr
\tabular{llllllll}{
 [col 1] series name|index \tab [col 2] line name \tab [col 3] Value
 \tab [col 4] P0 x \tab [col 5] P0 y \tab [col 6] P1 x \tab [col 7] P1 y
 \tab [col 8] Light effect \cr
 String|number \tab String|NA \tab num \tab x val \tab y val \tab x val
 \tab y val \tab TRUE|FALSE
}
Examples:
\itemize{
 \item E.g., both \code{t(c('male', NA, 'average', FALSE))} and \code{t(c(1, NA, 'average', FALSE))}
 refer to an average markline of the series 'male', only available for line,
 linesmooth, bar, scatter, bubble charts. 'lm' refers to linear regresson
 markline which is only available for scatters and bubbles.
 \code{t(c(1, NA, 'average', TRUE))} opens light effects of series 'male'. \cr
 \item In line, bar, k and scatter charts, 'P0 x','P0 y','P1 x','P1 y' are
 comprehended as coordinates. In map charts, these coordinates should be lattitudes
 and longitudes. \code{t(c('male', NA, 100, 0, 5, 100, 5, TRUE))} opens light effects of series 'male'.
}}

\item{markLinesmooth}{Used this instead of markLine for smooth marklines with
totally identical grammar.}

\item{markPoint}{Show markpoints, default to NULL.
The grammar is a data.frame or matrix with 4 or 6 columns:
Short form: \cr
\tabular{llll}{
 <col 1> series name|index \tab <col 2> Point name \tab <col 3> Point type \tab <col 4> Light effect \cr
 String or number \tab String|NA \tab min|max \tab TRUE|FALSE
 }
Full form: \cr
\tabular{llllll}{
 [col 1] series name|index \tab [col 2] Point name \tab [col 3] Value \tab [col 4] P x
 \tab [col 5] P y \tab [col 6] Light effect \cr
 String or number \tab String|NA \tab num \tab x val \tab y val \tab TRUE|FALSE
}
Examples: \cr
\itemize{
 \item E.g., both \code{t(c('male',NA,'min',FALSE))} and \code{t(c(1,NA,'min',FALSE))} refer
 to a min markpoint of the series 'male', only available for line, linesmooth, bar,
 scatter, bubble charts. \code{t(c('male',NA,'min',TRUE))} opens light effects of series 'male'.
 \item In line, bar, k and scatter charts, 'P x','P y',... are comprehended
 as coordinates. In map charts, these coordinates should be lattitudes and longitudes.
 \code{t(c('male',NA,100,0,5,TRUE))} opens light effects of series 'male'.
}}

\item{theme}{Default theme=\code{list(backgroundColor=color name/value,
borderColor=color name/value, borderWidth=1)}}

\item{...}{elipsis}
}
\value{
An echart object
}
\description{
A workaround function to generate echart objects.
}
\examples{
\dontrun{
echartR(iris, x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')

echartR(iris, x=~Sepal.Width, y=~Petal.Width, series=~Species)
}
}
\seealso{
\code{\link{ggthemes}} \code{\link{RColorBrewer}} \code{\link{echart}}
}

