% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{echartR}
\alias{echartR}
\title{A lattice-like workaround function to generate Echarts}
\usage{
echartR(data, x = NULL, y = x, z = NULL, series = NULL, weight = NULL,
  lat = NULL, lng = NULL, x1 = NULL, lat1 = NULL, lng1 = NULL,
  type = "auto", stack = FALSE, xlab = NULL, xAxis = list(lab = xlab,
  color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0), ylab = NULL,
  yAxis = list(lab = ylab, color = NULL, splitLine = TRUE, banded = FALSE,
  rotate = 0), xlab1 = NULL, xAxis1 = list(lab = xlab1, series = NULL,
  reverse = FALSE, color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0),
  ylab1 = NULL, yAxis1 = list(lab = ylab1, series = NULL, reverse = FALSE,
  color = NULL, splitLine = TRUE, banded = FALSE, rotate = 0), xyflip = FALSE,
  AxisAtZero = c(FALSE, TRUE), scale = TRUE, pos = list(title = 6, legend
  = 11, toolbox = 1, dataZoom = 6, dataRange = 8, roam = 2),
  theme = "macarons", ...)
}
\arguments{
\item{data}{data.frame. No defaults.}

\item{x}{x variable, only omitable for histograms, pie, ring and rose charts.}

\item{y}{y variable. If y is omitted, assign x to y.}

\item{z}{z variable, only accept data/time variable to open time axis. Optional.}

\item{series}{series variable for grouping.}

\item{weight}{weight variable, used in histogram, bubble, etc.}

\item{lat}{Lattitude variable, only for point/line-marking map.}

\item{lng}{Longitude variable, only for point/line-marking map.}

\item{x1}{Backup x variable, only for line-marking map, force andchord chart.}

\item{lat1}{Backup lattitude variable, only for line-marking map.}

\item{lng1}{Backup Longitude variable, only for line-marking map.}

\item{type}{options \code{'scatter', 'point', 'bubble', 'bar', 'line', 'map',
'k', 'pie', 'ring', 'rose','area', 'chord', 'force', 'tree', 'treemap',
'wordcloud', 'heatmap', 'histogram', 'funnel', 'pyramid', 'radar'} \cr
\itemize{
  \item If 'map' is chosen, the control option should be a vector of length 3:
  \code{c('map', mapType, markType)}. \code{mapType} could be either 'world' or 'china',
  of which simplified Chinese names are required for 'china'.
  When markType equals to 'area', the function colors polygons to show the effects;
  while equals to 'point', it ticks pins on the map; while equals to 'line',
  it ticks lines on the map.
  \item If 'line', 'area' are chosen, the control option should be a vector of
  length 2: \code{c(main type, type feature)}. \code{main type} is line or area,
  \code{type feature} is 'smooth'.
  \item If 'chord' is chosen, the control option should be a vector of
  length 2: \code{c('chord', 'riboon')}, which refers to chord chart with ribbon.
  \item If 'radar' is chosen, the control option should be a vector of
  length 2: \code{c('radar', 'fill')}, which refers to radar chart with color filled.
}}

\item{stack}{Default to FALSE (do not stack). Used in stacked column, bar, line
and area chart, etc.}

\item{xlab}{You can also omit xAxis, directly assign xAxis title.
xlab has a higher priority than xAxis[['lab']].}

\item{xAxis}{x Axis parameters in a list, default
\code{list(lab=NULL, color=NULL, splitLine=TRUE, banded=FALSE, rotate=0)}.}

\item{ylab}{You can also omit yAxis, directly assign yAxis title.
ylab has a higher priority than yAxis[['lab']].}

\item{yAxis}{parameters of y Axis. Refer to xAxis.}

\item{xlab1}{Title of secondary x-axis, with higher priority than lab in xAxis1 list,}

\item{xAxis1}{Secondary x axis, effecitve only if there are multiple series.
Written in a list, default \code{list(lab=NULL, series=NULL, reverse=FALSE,
color=NULL, splitLine=TRUE, banded=FALSE, rotate=0)}.
series set which series to put on secondary x-axis, which can be either a name
vector or an index vector, e.g., c(1,2) or c('male','female').
The data is fliped up when reverse is set TRUE.}

\item{ylab1}{Title of secondary y-axis, with higher priority than lab in yAxis1 list,}

\item{yAxis1}{Refer to xAxis1}

\item{xyflip}{Flip x, y-axies. Default FALSE.}

\item{AxisAtZero}{Axes (and sub axes) cross at zero. Default \code{c(FALSE, TRUE)}.}

\item{scale}{Rescale the axes based on min and max values. Default TRUE.}

\item{pos}{Position of image elements which are integers 1-12.
Title position default to 6 (o'clock), which means bottom middle.
Defaults to legend, toolbox, dataRange, dataZoom, roamController are
11, 1, 6, 8, 2, respectively. Default pos=list(title=6, legend=11, toolbox=1,
dataZoom=6, dataRange=8, roam=2)}

\item{...}{elipsis}
}
\value{
An echart object
}
\description{
A workaround function to generate echart objects.
}
\examples{
\dontrun{

echartR(iris, x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')

echartR(iris, x=~Sepal.Width, y=~Petal.Width, series=~Species)
}
}
\seealso{
\code{\link{ggthemes}} \code{\link{RColorBrewer}} \code{\link{echart}}
}

