% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eForce.R
\name{eForce}
\alias{eForce}
\title{Force network graph}
\usage{
eForce(networkMatrix, propertyDf = NULL, size = NULL, maxR = 25,
  minR = 15, density = 0.05, attractiveness = 1.2, showLabel = TRUE,
  theme = "default", title = NULL, subtitle = NULL, title.x = "center",
  title.y = "top", legend = TRUE, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  dataView = FALSE, readOnly = TRUE, mark = TRUE, dataZoom = FALSE,
  tooltip = TRUE, tooltip.trigger = "item", formatter = "",
  calculable = FALSE, xlab = NULL, ylab = NULL, opt = list())
}
\arguments{
\item{networkMatrix}{required, a symmetric matrix, each cell value indicates 
the weight of the two nodes and the 0 or NA cell would not be counted in. 
The matrix should have colnames or rownames.}

\item{propertyDf}{optional, dataframe which contain the metadata for the nodes. 
It could contain category, value and color columns. The colnames and rownames are required.}

\item{opt}{力导向图选项.}
}
\value{
The HTML code as a character string.
}
\description{
ECharts style 力导向图 graph visulize the social network matrix data.
}
\examples{
testData <- matrix(1:25, nrow=5) #测试中文
eForce(testData)
}
