% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/eFunnel.R
\name{eFunnel}
\alias{eFunnel}
\title{Funnel charts}
\usage{
eFunnel(dat, fix = T, opt = list())
}
\arguments{
\item{dat}{vector, names will be treated as labels}

\item{fix}{fix the max of data into 100 percents, default is true.}

\item{opt}{option of ECharts.}
}
\value{
an htmlwidgets object which can be rendered by Browser.
}
\description{
ECharts style Funnel charts.
}
\examples{
x = (1:5)*20
names(x) = LETTERS[1:5]
eFunnel(x,theme='macarons') +eTitle(title = "Funnel")
testData <- read.csv(system.file("examples", "testDataForMap.csv", package = "recharts"),stringsAsFactors=FALSE,encoding = "UTF-8")
eFunnel(head(testData), ~province, ~data2)
}

