% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePie.R
\name{ePie}
\alias{ePie}
\title{Pie charts}
\usage{
ePie(dat, namevar = NULL, valvar = NULL, size = NULL, type = c("pie",
  "rose"), roseType = c("radias", "area"), theme = "default",
  title = NULL, subtitle = NULL, title.x = "center", title.y = "top",
  legend = TRUE, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  dataView = TRUE, readOnly = FALSE, mark = TRUE, dataZoom = FALSE,
  magicType = FALSE, tooltip = TRUE, tooltip.trigger = "item",
  formatter = "", axis.scale = TRUE, xlab = FALSE, ylab = FALSE,
  calculable = TRUE, showLabel = TRUE, opt = list())
}
\arguments{
\item{dat}{vector, names will be treated as labels}

\item{opt}{option of ECharts.}
}
\value{
T	he HTML code as a character string.
}
\description{
ECharts style pie charts.
}
\examples{
x = runif(6)
names(x) = LETTERS[1:6]
ePie(x) + eTitle("test")
testData <- read.csv(system.file("examples", "testDataForMap.csv", package = "recharts"),stringsAsFactors=FALSE) 
ePie(head(testData), ~province, ~data2)
ePie(dat, ~namevar, ~valvar)
}

