% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ePoints.R
\name{ePoints}
\alias{ePoints}
\title{Points charts}
\usage{
ePoints(dat, xvar = NULL, yvar = NULL, series = NULL, size = NULL,
  power = 2, precision = 2, theme = "default", title = NULL,
  subtitle = NULL, title.x = "center", title.y = "top", legend = TRUE,
  legend.data = NULL, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  dataView = TRUE, readOnly = FALSE, mark = TRUE, dataZoom = TRUE,
  magicType = FALSE, tooltip = TRUE, tooltip.trigger = "item",
  formatter = "", axis.scale = TRUE, axis.line = TRUE,
  axis.tick = FALSE, axis.lable = TRUE, axis.splitLine = TRUE,
  axis.splitArea = FALSE, axis.boundaryGap = TRUE, xlab = TRUE,
  xlab.type = "value", xlab.data = NULL, xlab.position = "bottom",
  xlab.name = "", xlab.namePosition = "start", xlim = NULL, ylab = TRUE,
  ylab.type = "value", ylab.data = NULL, ylab.position = "left",
  ylab.name = "", ylab.namePosition = "start", ylim = NULL,
  calculable = FALSE, showLabel = TRUE, opt = list())
}
\arguments{
\item{dat}{data.frame, should have two column or three colume. 
If three, the third colume should be factor or character, it will be treated as category labels.}

\item{opt}{option of ECharts.}
}
\value{
The HTML code as a character string.
}
\description{
ECharts style scatter charts.
}
\examples{
ePoints(iris[,3:5], theme=2)
iris$Species <- as.character(iris$Species)
iris[1:20, "Species"] ="å°çº¢èŠ‘ù<U+3E31>"
ePoints(iris[,3:5], xvar=~Petal.Length, yvar=~Petal.Width, series=~Species, theme=2)

}

