% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eRadar.R
\name{eRadar}
\alias{eRadar}
\title{Create an HTML radar charts widget using the ECharts(version:3.2.2) library}
\usage{
eRadar(dat, xvar = NULL, yvar = NULL, series = NULL, size = NULL,
  ymin = vector(), ymax = vector(), theme = "default", title = NULL,
  subtitle = NULL, title.x = "center", title.y = "top", legend = TRUE,
  legend.data = NULL, legend.x = "left", legend.y = "top",
  legend.orient = "horizontal", toolbox = TRUE,
  toolbox.orient = "horizontal", toolbox.x = "right", toolbox.y = "top",
  dataView = TRUE, readOnly = FALSE, mark = TRUE, dataZoom = FALSE,
  magicType = FALSE, tooltip = TRUE, tooltip.trigger = "item",
  formatter = "", axis.scale = TRUE, xlab = FALSE, ylab = FALSE,
  calculable = TRUE, showLabel = TRUE, opt = list())
}
\arguments{
\item{dat}{a data object (a data frame or a factor array)}

\item{xvar, yvar}{objects of class "formula" (or one that can be coerced 
to that class):  x,y coordinates of the given data.frame colnames, e.g. 
\code{xvar = ~xAxisName}; \code{yvar = ~yAxisName}. xvar, yvar are needed for the 
data.frame data input.}

\item{series}{an "formula" object: Associates the levels of variable
with symbol color, e.g. \code{series = ~groupName}}

\item{size}{an array of html widget width and height(either numeric pixels 
or percentage could be accepted): e.g. size = c(1024, 768).}

\item{ymin, ymax}{a numeric array with the same length of unique(xvar), to set the
limitation for each polar axis limitation. if you want to set someone to be default
limitation, you can set the given array index to NULL, examples input:
\code{ymin = c(NULL, 1000, 2000, 3000)}}

\item{theme}{an object of theme name. see(\url{http://datatables.net/extensions/index}) for detail.
supported theme: \code{c("default", "vintage", "dark", "westeros", "essos", "wonderland", "walden",
"chalk", "infographic", "macarons", "roma", "shine", "purple-passion")}}

\item{title}{an overall title for the plot. you can modify title widget after chart has been
generated: Echart + eTitle(title = "your title.")}

\item{title.x, title.y}{the xy coordinates of main title, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify title widget after chart has been generated: 
\code{Echart + eTitle(title="main title", x = "left", y=10)}}

\item{legend}{logical whether the legend widget show or not, default is TRUE.
you can modify legend widget after chart has been generated, the legend position and 
legend orientation are available at present.
\code{Echart + eLegend(show = TRUE)}}

\item{legend.x, legend.y}{the xy coordinates of legend, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify legend widget after chart has been generated: 
\code{Echart + eLegend(x = "right", y="top")}}

\item{legend.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify legend widget after chart has been generated: 
\code{Echart + eLegend(orient = "vertical")}}

\item{toolbox}{logical whether the toolbox widget show or not, default is TRUE.
you can modify toolbox widget after chart has been generated, the toolbox position, toolbox 
element and toolbox orientation are available at present.  
\code{Echart + eToolbox(show = TRUE)}}

\item{toolbox.orient}{an element of c("horizontal", "vertical"), default is "horizontal"
you can modify toolbox widget after chart has been generated: 
\code{Echart + eToolbox(orient = "vertical")}}

\item{toolbox.x, toolbox.y}{the xy coordinates of toolbox, besides the excat exact pixels value, 
x accept c("left", "right", "center") and y accept c("top", "bottom", "center") as legal input. 
you can modify toolbox widget after chart has been generated: 
\code{Echart + eToolbox(x = "right", y="top")}}

\item{tooltip}{logical whether the tooltip widget for front-end interactive chart
show or not. default is TRUE. you can modify tooltip widget after chart has been generated, 
the tooltip trigger and tooltip formatter is available at present.  
\code{Echart + eTooltip(show = FALSE)}}

\item{tooltip.trigger}{an element of c("axis", "item"), default is "axis" for radar chart.
"axis" option for trigger will show all the information of mouse;
"item" option for tirgger will only show the given item information of mouse.
you can modify tooltip widget after chart has been generated: 
\code{Echart + eTooltip(trigger = "axis")}}

\item{calculable}{logical whether the front-end interactive chart will 
support the drag-recalculable feature.
the size and calculable option can be setted after radar chart has been 
generated through eOption: \code{Echart + eOption(calculable = TRUE)}}

\item{showLabel}{logical whether the region name label showed on chart.
default is TRUE, e.g. \code{Echart + eOption(showLabel = TRUE)}}

\item{dataview, mark, restore, dataZoom, magicType}{logical variable whether the dataview
, mark, restore, dataZoom or magicType tool in toolbox widget show or not, 
default is TRUE. you can modify toolbox widget after chart has been generated, 
the toolbox position, toolbox element and toolbox orientation are available at present.  
\code{Echart + eToolbox(dataView = FALSE)}}

\item{tooltip.formatter}{the information formatter for tooltip widget, 
default is "<a>:<b><c>" for radar chart.
you can modify tooltip widget after chart has been generated: 
\code{Echart + eTooltip(formatter = "<a><b>:<c>")}}
}
\description{
This function creates an HTML widget to display data.frame using the JavaScript library ECharts3.
}
\note{
You are recommended to use lazyPlot function for interactive chart
  option set through "shiny" server.
}
\examples{
require(plyr)
dat = ddply(iris, .(Species), colwise(mean))
rownames(dat) = dat[,1]
dat = dat[, -1]
dat
eRadar(dat)
df2 <- data.frame(
 saleNum=c(10,20,30,40,50,60,70,15,25,35,45,55,65,75,25,35,45,55,65,75,85),
	seller=c(rep("Yellow",7), rep("Red",7), rep("White",7)),
weekDay = c(rep(c("Mon","Tue","Wed","Thu","Fri","Sat","Sun"),3))
)
dat <- df2
xvar=~weekDay; yvar= ~saleNum; series=~seller
eRadar(df2, ~weekDay, ~saleNum, ~seller)
}
