% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chartOpt.R
\name{eTitleSet}
\alias{eTitleSet}
\title{recharts tilte Set fucntion}
\usage{
eTitleSet(chart, ...)
}
\arguments{
\item{...}{default parameter.}

\item{x}{recharts print object.}

\item{tag}{whether print the recharts object to browser or string.}
}
\value{
The HTML code as a character string.
}
\description{
An shell function for printing of the recharts object.
}
