% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setGrid}
\alias{setGrid}
\title{Set \code{grid} of Echarts (Only for Charts Of Cartesian Coordinates System)}
\usage{
setGrid(chart, x = 80, y = 60, x2 = 80, y2 = 60, width = NULL,
  height = NULL, bgColor = NULL, borderColor = NULL, borderWidth = 1,
  ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{x}{x coordinate of the left upper point of the plot area. Default 80px.}

\item{y}{y coordinate of the left upper point of the plot area. Default 60px.}

\item{x2}{x coordinate of the right lower point of the plot area. Default 80px.}

\item{y2}{y coordinate of the right lower point of the plot area. Default 60px.}

\item{width}{Width of the plot area. Default NULL (automatically configured)}

\item{height}{Height of the plot area. Default NULL (automatically configured)}

\item{bgColor}{background color of plot area. Default transparent ('rgba(0,0,0,0)').}

\item{borderColor}{border color of the plot area. Default '#ccc'.}

\item{borderWidth}{border width of the plot area. Default 0px (not shown).}

\item{...}{Elipsis.}
}
\value{
A modified echarts object
}
\description{
When an echart object is generated, you can modify it by setting grid using
\code{\link{\%>\%}}.
It is recommended to put \code{setGrid} at the end.
Only applicable for \code{scatter, point, bubble, line, area, bar, histogram}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species)
g \%>\% setGrid(x=40, y=40, x2=70, y2=30, bgColor='gray90')
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~grid}
}

