% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setGrid}
\alias{setGrid}
\alias{tunePosDataRange}
\alias{tunePosDataZoom}
\alias{tunePosLegend}
\alias{tunePosRoam}
\alias{tunePosTimeline}
\alias{tunePosTitle}
\alias{tunePosToolbox}
\title{Set \code{grid} of Echarts Widgts And Pane}
\usage{
setGrid(chart, x = 80, y = 60, x2 = 80, y2 = 60, width = NULL,
  height = NULL, bgColor = NULL, borderColor = NULL, borderWidth = 1,
  widget = c("pane", "timeline", "legend", "title", "dataZoom", "dataRange",
  "toolbox", "roamController"), ...)

tunePosTitle(chart, x, y, ...)

tunePosLegend(chart, x, y, ...)

tunePosDataZoom(chart, x, y, ...)

tunePosDataRange(chart, x, y, ...)

tunePosTimeline(chart, x, y, x2, y2, ...)

tunePosToolbox(chart, x, y, ...)

tunePosRoam(chart, x, y, ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{x}{x coordinate of the left upper point of the plot area. Default 80px.}

\item{y}{y coordinate of the left upper point of the plot area. Default 60px.}

\item{x2}{x coordinate of the right lower point of the plot area. Default 80px.}

\item{y2}{y coordinate of the right lower point of the plot area. Default 60px.}

\item{width}{Width of the plot area. Default NULL (automatically configured)}

\item{height}{Height of the plot area. Default NULL (automatically configured)}

\item{bgColor}{background color of plot area. Default transparent ('rgba(0,0,0,0)').}

\item{borderColor}{border color of the plot area. Default '#ccc'.}

\item{borderWidth}{border width of the plot area. Default 0px (not shown).}

\item{widget}{Widget name to set. Could be \code{c('pane', 'timeline', 'legend',
'title', 'dataZoom', 'dataRange', 'toolbox', 'roamController')}.
\describe{
\item{pane}{the area pane, takes in all the parameters}
\item{timeline}{timeline widget, only use \code{x, y, x2, y2}}
\item{legend, title, dataZoom, dataRange, toolbox, roamController}{other widgets,
only use \code{x, y}}
}}

\item{...}{Elipsis.}
}
\value{
A modified echarts object
}
\description{
When an echart object is generated, you can modify it by setting grid using
\code{\link{\%>\%}}. \cr
\strong{It is recommended to put \code{setGrid} at the end of the piped command.} \cr
When used for 'pane', it is only applicable for \code{scatter, point, bubble,
line, area, bar, histogram}. When used for 'timeline', it only take in params
\code{x, y, x2, y2}. When used for 'legend', 'title', 'dataZoom', 'dataRange',
 'toolbox', 'roamController', it only takes in params \code{x, y}.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species)
g \%>\% setGrid(x=40, y=40, x2=70, y2=30, bgColor='gray90')
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/option.html#title~grid}
}

