% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{setTheme}
\alias{setTheme}
\title{Set Theme of Echarts}
\usage{
setTheme(chart, theme = c("default", "macarons", "infographic", "blue",
  "dark", "gray", "green", "helianthus", "macarons2", "mint", "red", "roma",
  "sakura", "shine", "vintage"), ...)
}
\arguments{
\item{chart}{\code{echarts} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{theme}{\itemize{
\item Pre-installed themes: \code{'default', 'macarons', 'infographic', 'blue', 'dark', 'gray',
'green', 'helianthus', 'macarons2', 'mint', 'red', 'roma', 'sakura', 'shine', 'vintage'} \cr
\item User-defined themes: e.g., \code{list(color=c('#ff7f50','#87cefa','#da70d6','#32cd32','#6495ed'), \cr
title=list(x='left', y='top'), legend=list(orient='horizontal'))}
}}

\item{...}{Elipsis}
}
\value{
A modified echarts object
}
\description{
Set the theme of Echarts. You can use a pre-installed theme or define you own themes. \cr
When an echart object is generated, you can modify it by setting theme using
\code{\link{\%>\%}}.
}
\section{Self-defined Themes (UDT)}{

You can design your prefered themes using \url{http://echarts.baidu.com/echarts2/doc/example/themeDesigner.html}.\cr
\cr The UDTs in \code{echarts} are JS objects. You can write it in a list form and \code{setTheme}
will parse it into JSON string for process. \cr \cr
A typical theme JS looks like: \cr
\code{{
    color: ['#ff7f50','#87cefa','#da70d6','#32cd32','#6495ed'], \cr
    title: {x: 'left', y: 'top'}, \cr
    legend: {orient: 'horizontal'}
}} \cr \cr
You should write in the format below: \cr
\code{list(color=c('#ff7f50','#87cefa','#da70d6','#32cd32','#6495ed'), \cr
title=list(x='left', y='top'), legend=list(orient='horizontal'))}
}
\examples{
\dontrun{
g <- echartR(iris, Sepal.Width, Petal.Width, series=Species)

## pre-installed themes
g \%>\% setTheme("infographic")
g \%>\% setTheme("blue")
g \%>\% setTheme("dark")
g \%>\% setTheme("gray")
g \%>\% setTheme("green")
g \%>\% setTheme("helianthus")
g \%>\% setTheme("macarons2")
g \%>\% setTheme("mint")
g \%>\% setTheme("red")
g \%>\% setTheme("sakura")
g \%>\% setTheme("shine")
g \%>\% setTheme("vintage")

## self-defined themes
theme <- list(color=c("#7AC143", "#7D3F98", "#F47721", "#D20962", "#00A78E",
              "#00BCE4", "#B8D936", "#EE3D94"), backgroundColor="#fef8ef")
g \%>\% setTheme(theme)
}
}
\references{
\url{http://echarts.baidu.com/echarts2/doc/example/theme.html}
}

