% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{setTitle}
\alias{setTitle}
\title{Set Title And Subtitle of Echarts}
\usage{
setTitle(chart, title = NULL, link = NULL, subtitle = NULL,
  sublink = NULL, pos = 6, ...)
}
\arguments{
\item{chart}{\code{echart} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{title}{text of the title}

\item{link}{link of the title}

\item{subtitle}{text of the subtitle}

\item{sublink}{link of the subtitle}

\item{pos}{the clock-position of title (and subtitle), refert to \code{\link{rechart:::vecPos}}}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting title and
subtitles using %>%.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setTitle(title='Iris', subtitle='by: Fisher')
}
}

