% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/echartR.R
\name{setToolbox}
\alias{setToolbox}
\title{Set Toolbox of Echarts}
\usage{
setToolbox(chart, show = TRUE, language = "cn", controls = c("mark",
  "dataZoom", "dataView", "magicType", "restore", "saveAsImage"), pos = 1,
  ...)
}
\arguments{
\item{chart}{\code{echart} object generated by \code{\link{echart}} or \code{\link{echartR}}}

\item{show}{logical. Show the toolbox if TRUE.}

\item{language}{'cn' or 'en', the language of the toolbox tooltips.}

\item{controls}{which widgets to show. Default \code{'mark', 'dataZoom', 'dataView', 'magicType',
'restore', 'saveAsImage'}.}

\item{pos}{the clock-position of toolbox, refert to \code{\link{vecPos}}}

\item{...}{elipsis}
}
\value{
A modified echart object
}
\description{
When an echart object is generated, you can modify it by setting toolbox using \%>\%.
}
\examples{
\dontrun{
g <- iris \%>\% echartR(x=Sepal.Width, y=Petal.Width, series=Species, type='scatter')
g \%>\% setToolbox(TRUE, 'en')
}
}

