% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optSetUtils.R
\name{tilteSet}
\alias{tilteSet}
\title{recharts tilteSet fucntion}
\usage{
tilteSet(title = NULL, link = NULL, subtitle = NULL, sublink = NULL,
  title.x = "left", title.y = "top")
}
\arguments{
\item{x}{recharts print object.}

\item{tag}{whether print the recharts object to browser or string.}

\item{...}{default parameter.}
}
\value{
The HTML code as a character string.
}
\description{
An shell function for printing of the recharts object.
}
